# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Voice-first Urdu personal finance assistant backend (Uplift AI STT/TTS + LLM intent extraction + finance engine).
 * OpenAPI spec version: 2.0.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
}

export interface ApiError {
  error: string;
}

export interface TranscriptionResult {
  /** Urdu transcription of the audio */
  text: string;
}

export interface ExtractRequest {
  /** @minLength 1 */
  text: string;
}

export type ExtractResultMode = typeof ExtractResultMode[keyof typeof ExtractResultMode];


export const ExtractResultMode = {
  transaction: 'transaction',
  query: 'query',
  unknown: 'unknown',
} as const;

/**
 * @nullable
 */
export type ExtractResultType = typeof ExtractResultType[keyof typeof ExtractResultType] | null;


export const ExtractResultType = {
  expense: 'expense',
  income: 'income',
  given: 'given',
  received: 'received',
} as const;

/**
 * @nullable
 */
export type ExtractResultQueryType = typeof ExtractResultQueryType[keyof typeof ExtractResultQueryType] | null;


export const ExtractResultQueryType = {
  total_expenses: 'total_expenses',
  total_income: 'total_income',
  person_given: 'person_given',
  person_received: 'person_received',
  person_balance: 'person_balance',
  today_summary: 'today_summary',
  monthly_summary: 'monthly_summary',
  category_summary: 'category_summary',
  recent_transactions: 'recent_transactions',
} as const;

/**
 * @nullable
 */
export type ExtractResultPeriod = typeof ExtractResultPeriod[keyof typeof ExtractResultPeriod] | null;


export const ExtractResultPeriod = {
  today: 'today',
  this_week: 'this_week',
  this_month: 'this_month',
  all_time: 'all_time',
} as const;

/**
 * Structured intent. mode=transaction fills type/amount/person/category/description; mode=query fills query_type/period/person/category; mode=unknown has no other fields.
 */
export interface ExtractResult {
  mode: ExtractResultMode;
  /** @nullable */
  type?: ExtractResultType;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  person?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  query_type?: ExtractResultQueryType;
  /** @nullable */
  period?: ExtractResultPeriod;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  expense: 'expense',
  income: 'income',
  given: 'given',
  received: 'received',
} as const;

export interface TransactionInput {
  /** Rupee amount (positive) */
  amount: number;
  type: TransactionInputType;
  /** @nullable */
  person?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  expense: 'expense',
  income: 'income',
  given: 'given',
  received: 'received',
} as const;

export interface Transaction {
  id: string;
  amount: number;
  type: TransactionType;
  /** @nullable */
  person: string | null;
  /** @nullable */
  category: string | null;
  /** @nullable */
  description: string | null;
  timestamp: string;
}

export interface CreateTransactionResult {
  transaction: Transaction;
  /** Deterministic Urdu confirmation sentence to speak aloud */
  responseText: string;
}

export interface Summary {
  income: number;
  expenses: number;
  given: number;
  received: number;
}

export interface TransactionsList {
  transactions: Transaction[];
  summary: Summary;
}

export interface PersonLedger {
  person: string;
  given: number;
  received: number;
  /** given - received (what the person still owes the user) */
  balance: number;
  transactions: Transaction[];
}

export type QueryRequestQueryType = typeof QueryRequestQueryType[keyof typeof QueryRequestQueryType];


export const QueryRequestQueryType = {
  total_expenses: 'total_expenses',
  total_income: 'total_income',
  person_given: 'person_given',
  person_received: 'person_received',
  person_balance: 'person_balance',
  today_summary: 'today_summary',
  monthly_summary: 'monthly_summary',
  category_summary: 'category_summary',
  recent_transactions: 'recent_transactions',
} as const;

/**
 * @nullable
 */
export type QueryRequestPeriod = typeof QueryRequestPeriod[keyof typeof QueryRequestPeriod] | null;


export const QueryRequestPeriod = {
  today: 'today',
  this_week: 'this_week',
  this_month: 'this_month',
  all_time: 'all_time',
} as const;

export interface QueryRequest {
  query_type: QueryRequestQueryType;
  /** @nullable */
  period?: QueryRequestPeriod;
  /** @nullable */
  person?: string | null;
  /** @nullable */
  category?: string | null;
}

/**
 * Raw calculation result from the finance engine
 */
export type QueryResultResult = { [key: string]: unknown };

export interface QueryResult {
  /** Raw calculation result from the finance engine */
  result: QueryResultResult;
  /** Deterministic Urdu sentence to speak aloud */
  responseText: string;
}

export interface SpeakRequest {
  /** @minLength 1 */
  text: string;
}

export interface SpeakResult {
  /** Base64-encoded MP3 audio */
  audio: string;
  /** e.g. audio/mpeg */
  contentType: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ApiError;

/**
 * Not found
 */
export type NotFoundResponse = ApiError;

/**
 * Upstream AI service failed
 */
export type UpstreamErrorResponse = ApiError;

export type TranscribeAudioBody = {
  /** Recorded audio file (m4a/webm/wav) */
  audio: Blob;
};