# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Awaz Khata API
  description: Voice-first Urdu personal finance assistant backend (Uplift AI STT/TTS + LLM intent extraction + finance engine).
  version: 2.0.0

servers:
  - url: /api

paths:
  /healthz:
    get:
      operationId: healthCheck
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /voice/transcribe:
    post:
      operationId: transcribeAudio
      summary: Transcribe an Urdu voice recording to text (Uplift AI STT)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: Recorded audio file (m4a/webm/wav)
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"

  /voice/extract:
    post:
      operationId: extractIntent
      summary: Extract a structured financial intent from transcribed text (LLM)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractRequest"
      responses:
        "200":
          description: Structured intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"

  /voice/speak:
    post:
      operationId: speakText
      summary: Convert Urdu text to speech (Uplift AI TTS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakRequest"
      responses:
        "200":
          description: Base64 MP3 audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"

  /transactions:
    post:
      operationId: createTransaction
      summary: Save a financial transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Saved transaction plus a spoken Urdu confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTransactionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listTransactions
      summary: List all transactions with overall summary
      responses:
        "200":
          description: All transactions (newest first) and totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsList"

  /transactions/person/{person}:
    get:
      operationId: getPersonLedger
      summary: Get one person's given/received totals and balance (case-insensitive)
      parameters:
        - name: person
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Person ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonLedger"
        "404":
          $ref: "#/components/responses/NotFound"

  /query:
    post:
      operationId: runQuery
      summary: Run a structured financial query through the finance engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Query result with a spoken Urdu response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
        "400":
          $ref: "#/components/responses/BadRequest"

components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    UpstreamError:
      description: Upstream AI service failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    TranscriptionResult:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: Urdu transcription of the audio

    ExtractRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1

    ExtractResult:
      type: object
      description: >
        Structured intent. mode=transaction fills type/amount/person/category/description;
        mode=query fills query_type/period/person/category; mode=unknown has no other fields.
      required: [mode]
      properties:
        mode:
          type: string
          enum: [transaction, query, unknown]
        type:
          type: string
          enum: [expense, income, given, received]
          nullable: true
        amount:
          type: number
          nullable: true
        person:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        query_type:
          type: string
          enum:
            - total_expenses
            - total_income
            - person_given
            - person_received
            - person_balance
            - today_summary
            - monthly_summary
            - category_summary
            - recent_transactions
          nullable: true
        period:
          type: string
          enum: [today, this_week, this_month, all_time]
          nullable: true

    TransactionInput:
      type: object
      required: [amount, type]
      properties:
        amount:
          type: number
          description: Rupee amount (positive)
        type:
          type: string
          enum: [expense, income, given, received]
        person:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true

    Transaction:
      type: object
      required: [id, amount, type, person, category, description, timestamp]
      properties:
        id:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [expense, income, given, received]
        person:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    CreateTransactionResult:
      type: object
      required: [transaction, responseText]
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        responseText:
          type: string
          description: Deterministic Urdu confirmation sentence to speak aloud

    Summary:
      type: object
      required: [income, expenses, given, received]
      properties:
        income:
          type: number
        expenses:
          type: number
        given:
          type: number
        received:
          type: number

    TransactionsList:
      type: object
      required: [transactions, summary]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        summary:
          $ref: "#/components/schemas/Summary"

    PersonLedger:
      type: object
      required: [person, given, received, balance, transactions]
      properties:
        person:
          type: string
        given:
          type: number
        received:
          type: number
        balance:
          type: number
          description: given - received (what the person still owes the user)
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"

    QueryRequest:
      type: object
      required: [query_type]
      properties:
        query_type:
          type: string
          enum:
            - total_expenses
            - total_income
            - person_given
            - person_received
            - person_balance
            - today_summary
            - monthly_summary
            - category_summary
            - recent_transactions
        period:
          type: string
          enum: [today, this_week, this_month, all_time]
          nullable: true
        person:
          type: string
          nullable: true
        category:
          type: string
          nullable: true

    QueryResult:
      type: object
      required: [result, responseText]
      properties:
        result:
          type: object
          additionalProperties: true
          description: Raw calculation result from the finance engine
        responseText:
          type: string
          description: Deterministic Urdu sentence to speak aloud

    SpeakRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1

    SpeakResult:
      type: object
      required: [audio, contentType]
      properties:
        audio:
          type: string
          description: Base64-encoded MP3 audio
        contentType:
          type: string
          description: e.g. audio/mpeg