# lib/api-zod/src/generated/types/speakResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Voice-first Urdu personal finance assistant backend (Uplift AI STT/TTS + LLM intent extraction + finance engine).
 * OpenAPI spec version: 2.0.0
 */

export interface SpeakResult {
  /** Base64-encoded MP3 audio */
  audio: string;
  /** e.g. audio/mpeg */
  contentType: string;
}