# lib/api-zod/src/generated/types/transactionsList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Voice-first Urdu personal finance assistant backend (Uplift AI STT/TTS + LLM intent extraction + finance engine).
 * OpenAPI spec version: 2.0.0
 */
import type { Summary } from './summary';
import type { Transaction } from './transaction';

export interface TransactionsList {
  transactions: Transaction[];
  summary: Summary;
}